Add Gradle version compatibility integration tests
Test the plugin against Gradle 9.1.0 and 9.4.1 using
TestParameterInjector to catch version-specific breakage in
internal Gradle APIs without significantly increasing build times.

Also extract the shared runFixture helper from PaparazziPluginTest
into a top-level extension function in TestUtils.kt.
