+++ paparazzi-gradle-plugin/src/test/java/app/cash/paparazzi/gradle/PaparazziPluginTest.kt
@@ -1737,37 +1737,6 @@ class PaparazziPluginTest {
 
   private fun File.loadConfig() = source().buffer().use { CONFIG_ADAPTER.fromJson(it)!! }
 
-  private fun GradleRunner.runFixture(projectRoot: File, action: GradleRunner.() -> BuildResult): BuildResult {
-    val settings = File(projectRoot, "settings.gradle")
-    val gradleProperties = File(projectRoot, "gradle.properties")
-    var generatedSettings = false
-    var generatedGradleProperties = false
-
-    return try {
-      if (!settings.exists()) {
-        settings.createNewFile()
-        settings.writeText("apply from: \"../test.settings.gradle\"")
-        generatedSettings = true
-      }
-
-      if (!gradleProperties.exists()) {
-        gradleProperties.createNewFile()
-        gradleProperties.writeText(
-          """
-            |android.useAndroidX=true
-            |android.dependencyResolutionAtConfigurationTime.disallow=true
-          """.trimMargin()
-        )
-        generatedGradleProperties = true
-      }
-
-      withProjectDir(projectRoot).action()
-    } finally {
-      if (generatedSettings) settings.delete()
-      if (generatedGradleProperties) gradleProperties.delete()
-    }
-  }
-
   private fun File.registerForDeletionOnExit() = apply { filesToDelete += this }
 
   private fun File.listFilesSorted() = listFiles()?.sortedBy { it.lastModified() }
